+++ CHANGES.md
@@ -3,7 +3,9 @@
 ## 2.2.10 (...)
 - make sure the IPv6 pool always starts at `::1000` which was the default on
   OpenVPN 2.4, but changed to `::2` in OpenVPN 2.5 for `/112` prefixes
-  
+- fix bug in `IP::getFirstHost()` that would not start from "network" address, 
+  but from IP address
+
 ## 2.2.9 (2021-08-02)
 - also provide "originating" client IP to the server API for logging